fixing db issues and adding create_pot api

+++ expense-tracker/expense-tracker-db/src/splits.rs
@@ -1,16 +1,19 @@
 pub mod splits {
-    use diesel::{Associations, Insertable, PgConnection, Queryable, Selectable};
+    use diesel::{Associations, Insertable, Queryable, Selectable};
     use serde::{Deserialize, Serialize};
+    use crate::expenses::expenses::Expense;
+    use crate::schema::expense_splits;
 
     /// This struct represents a split which is in turn part of an Expense
     /// but related to a user. The user is the one owing money the owner of
     /// the expense.
     #[derive(Serialize, Selectable, Queryable, Associations)]
     #[diesel(belongs_to(Expense))]
+    #[diesel(table_name = expense_splits)]
     pub struct Split {
         expense_id : i32,
         user_id : i32,
-        amount : i32,
+        amount : f64,
         is_paid: bool
     }
 
@@ -19,7 +22,7 @@ pub mod splits {
         pub fn new(
             expense_id : i32,
             user_id : i32,
-            amount : i32,
+            amount : f64,
             is_paid: bool
         ) -> Self {
             Split {
@@ -41,7 +44,7 @@ pub mod splits {
         }
 
         /// Getter for amount.
-        pub fn amount(&self) -> i32 {
+        pub fn amount(&self) -> f64 {
             self.amount
         }
 
@@ -53,11 +56,11 @@ pub mod splits {
 
     /// Struct used to create a new Split in the db.
     #[derive(Deserialize, Insertable)]
-    #[diesel(table_name = splits)]
+    #[diesel(table_name = expense_splits)]
     pub struct NewSplit {
         expense_id : i32,
         user_id : i32,
-        amount : i32,
+        amount : f64,
         is_paid: bool
     }
 }
